+++ module/data-models/item.spell.js
@@ -1,0 +1,10 @@
+import { WeaponDataModel } from './item.weapon.js';
+
+export class SpellDataModel extends WeaponDataModel {
+  static get defaults() {
+    return {
+      ...super.defaults,
+      ignoreSlots: true,
+    };
+  }
+}